# Add new trimming approach: Highest Quality Subread