--- a/src/main.rs
+++ b/src/main.rs
@@ -8,6 +8,7 @@ use std::path::PathBuf;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
+mod trimmers;
 mod utils;
 use utils::file_reader;