Hi @wdecoster,
I've used Chopper a couple of times and found it to be a very useful and easy-to-use trimming tool. I wanted to contribute my small grain of sand. I apologize, as I initially started with a simple idea, but along the way I ended up making additional changes, which might make it a bit harder to review everything in detail. Here’s a brief overview of my contribution:

## What's new?

A new trimming strategy called **Highest Quality Subread** was added. It is based on the algorithm presented by [Phred](http://www.phrap.org/phredphrap/phred.html), which identifies the highest-quality subread based on a cutoff value (`--cutoff`).

## How does the algorithm work?

The behavior of this new trimming approach can be summarized in four steps:

1. For each base in the read, the Phred quality score is converted into an error probability.
2. For each base, a score is calculated as the difference between the `--cutoff` value and the error probability; bases with a negative score are considered to have insufficient quality.
3. The subsequence with the highest cumulative score is identified using Kadane’s algorithm.
4. The indices of the highest-quality subread are returned, if one is found.

## Types of changes

### New feature

- A new trimming approach was implemented, and the `TrimStrategy` trait was created to standardize the logic across different strategies, making the system easier to extend and maintain.
- The `--trim-approach` flag was added to allow users to select one of the available trimming strategies, while ensuring only one strategy is applied at a time.
- The `--cutoff` flag was added to define a quality threshold for strategies that require it (e.g., `trim-by-quality` and `best-subread`).

### Additional implementations

- `parse_gc_value` function was added to parse GC content parameters and validate that their values fall within the `[0, 1]` range.
- `parse_usize_or_inf` function was added to parse the maximum read length, allowing the terminal to display `INF` (instead of `usize::MAX`) as default value.
- The `build_trimming_approach` function was added to construct the appropriate trimming strategy and validate that all required parameters (`--cutoff`, `--headcrop`, `--tailcrop`) have been provided for the selected strategy.

### Refactoring

- The `--trim` flag was removed and replaced by `--cutoff`, which must now be used together with the `--trim-approach trim-by-quality` flag.
- The new `--trim-approach` flag allows selecting one of three strategies: fixed-length trimming (`fixed-crop`), quality-based trimming (`trim-by-quality`), and best subread extraction (`best-subread`).
- CLI argument presentation was reorganized and grouped into three categories in the help output: **Filtering options**, **Trimming options**, and **Setup options**.
- Trimming logic was extracted from the `write_record` function and encapsulated into dedicated structs implementing the `TrimStrategy` trait.
- The `write_record` function was simplified to focus solely on writing reads to output.
- The `filter` function was refactored to improve readability and modularity.
- All tests using the `Cli` structure were updated accordingly.

### Bugfix

- Validation for `--gcmin` and `--gcmax` was added to the `filter` function when the `--inverse` flag is used.

## Testing

- All existing tests were verified and passed successfully.
- The tool was tested on a Linux system (Debian 12) and compared against version `0.10.0` installed via Conda.
- Multiple FASTQ files were used to confirm consistent results. Identical output was observed in all cases, except when using the `--inverse` flag with `--gcmin` and/or `--gcmax`, where the new logic now takes GC limits into account when `--inverse` flag is set.

## Benchmark

To evaluate performance, `chopper v0.10.0` (installed via Conda) was compared with the new version using a 8.9 GB FASTQ file. The following parameters were used:

```bash
# Old version
chopper -t <THREADS> --trim 25 --mingc 0.2 --maxgc 0.8 -q 25 -i <TEST_FILE>

# New version
<PATH_TO_CHOPPER_RELEASE> -t <THREADS> --trim-approach trim-by-quality --cutoff 25 --mingc 0.2 --maxgc 0.8 -q 25 -i <TEST_FILE>
```

Each version was executed 20 times using 1 and 4 threads, and the average execution time (in seconds) was recorded. The table below shows similar performance between both versions, with the new version being slightly faster on average.

| Version             | Avg time (1 thread) | Avg time (4 threads) |
|---------------------|---------------------|----------------------|
| chopper_v0.10.0     | 95.1797 s           | 39.2535 s            |
| chopper_new_version | 93.4072 s           | 39.2436 s            |