@wdecoster 2025-07-02
Hi @MillarCD,

Wow! Thanks for contributing! I will need some time to go through this. Looks like a great addition, while making things a bit more complicated. I haven't worked with Traits much, so that's nice too. Also, thanks for the bug fix of the GC content.

One thing that I am conflicted about is how you named the strategy. A 'subread' is terminology used by PacBio for individual passes of the polymerase around the fragment, before merging them all in a ccs read. Can you think of an alternative?

Best,
Wouter