@MillarCD 2025-08-03
> It came up with two spelling fixes and one potential underflow, which appears a valid concern to me, do you agree?

Yes, it was a small detail that was easy to fix but with a significant impact.